requirements.txt
requests==2.22.0
pandas==0.25.0
beautifulsoup4==4.8.1
python-dotenv==0.10.3
tweepy==3.8.0
